Boom WP Theme With Persistent Audio Player
This is kinda crazy but I’ve put the Wolf Player at the top of the page
just inside the HTML Body tag. It’s not the most peformant thing in the
world but I’m thinking we can just style the player to be really small,
like 44 pixels in height so people can use it on their phones a bit
easier. It’s not something that needs to be huge. Just realized I don’t
know what it will look like where there’s a huge playlist.